pkg/objc: correct objc_msgSend call on x86_64 depending on return type
On x86_64, we have to use specialized forms of `objc_msgSend` depending
on the return type (and the byte size of the return type). If we don't
do this, the calling convention is wrong and we'll get a segfault when
objc_msgSend tries to read a register that is [rightly] not set.

On aarch64, we do not have this issue because the calling convention is
different and objc_msgSend handles all cases.