config: introduce config-default-files to skip loading default files
This is a CLI-only config. This is the first such config. I am only
pointing this out since it is a new pattern for Ghostty.

By specifying `--config-default-files=false`, Ghostty will discard any
configuration set in the default files it loads. This makes running
Ghostty from the CLI easier if you want to "reset" the configuration to
the "factory defaults."